requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
mistralai
pdfplumber
python-multipart
